Allow struct-wide defaults with fxstruct attribute
I.e. if `into` is specified there it becomes the default for all fields.

Along the lines improved handling of attribute arguments switching to
function-like syntax on all of them except for `default`. I.e. where
previously it was `reader = off` it is now `reader(off)`. Besides, it
allows for `reader("reader_name")` and, even, for `reader("reader_name",
off)` – mostly, for debugging purposes when it has to be temporarily
disabled.

`builder` argument can do even more complex things:

```
    builder(
        attributes(allow(dead_code)),
        attributes_impl( allow(dead_code) )
    )
)]
```