feat: Generate errors for useless sub-arguments
Say, we have a field with a `serde` declaration:

```rust
foo: i32,
```

But shadow struct generation is a responsibility of struct-level code.
So, the `shadow_name` argument is useless here and should be reported as
an error.

Note: will be re-implemented as a warning as soon as the diagnostics
become available in stable.