feat: Implement simplified `lazy` implementation for non-lock fields
Annotation: It is not always necessary to lock-protect lazy fields; this
is especially true for read-only ones. Therefore, unless the `lazy`
attribute is accompanied by the `lock` or `inner_mut` attributes (which
are just aliases in the `sync` and `async` modes), lazy evaluation will
be implemented using `OnceCell` from either `once_cell` or `tokio` for
the `sync` and `async` contexts, respectively.