feat: Allow optional alternative async-lock instead of tokio
Annotation: It’s been long planned to relax the crate’s dependency on
`tokio`. You can now use the `async-lock` crate to implement `RwLock`
and `OnceCell`. This change also updates the feature flags: the `async`
feature must be paired with either `tokio-backend` or
`async-lock-backend`. Alternatively, enable async support and select an
implementation in one step with the umbrella flags `async-tokio` or
`async-lock`.

`tokio` remains in the requirements for tests.