Make FXSerdeHelper aware of its struct/field level context
According to the separation of concerns principle, delegate the
verification of acceptable subarguments to the helper itself.
The implementation also involved extending the validate_no_macro_args
macro with an additional syntax and giving it to a more accurate
name `validate_no_subargs_at_level` for better clarity.

The builder helper got involved as a good fit for the new macro syntax.

Now unnecessary checks are removed from the field receiver
implementation.