Make README generation more flexible
None of the README generators that were considered allow for extending
template context with additional data. Besides, the one that was used
before (`cargo-readme`) doesn't handle links. So, it has been replaced
with `cargo-doc2readme` in the first place.

Then, a two-layered README transformation was introduced, where the
first layer is the actual generation from the source code using
`cargo-doc2readme`.  The second layer is expansion using the `minijinja`
templating engine with a custom-generated JSON context, which includes
`git.branch` and `readme` template variables where the `readme` variable
is the output of the first layer.